feat: add ja.senmanga
