+++ sources/ja.senmanga/res/filters.json
@@ -2,21 +2,22 @@
 	{
 		"type": "sort",
 		"id": "order",
+		"title": "並び替え",
 		"canAscend": false,
-		"options": ["Popular", "Title", "Updated", "Rating"]
+		"options": ["人気", "タイトル", "更新順", "評価"]
 	},
 	{
 		"type": "select",
 		"id": "status",
-		"title": "Status",
-		"options": ["Any", "Ongoing", "Completed", "Hiatus", "Cancelled"],
+		"title": "状態",
+		"options": ["全て", "連載中", "完結", "休載", "打ち切り"],
 		"ids": ["", "Ongoing", "Completed", "Hiatus", "Cancelled"]
 	},
 	{
 		"type": "select",
 		"id": "type",
-		"title": "Type",
-		"options": ["Any", "Manga", "Manhwa", "Manhua"],
+		"title": "タイプ",
+		"options": ["全て", "Manga", "Manhwa", "Manhua"],
 		"ids": ["", "Manga", "Manhwa", "Manhua"]
 	}
 ]
