+++ sources/ja.senmanga/res/source.json
@@ -5,6 +5,7 @@
 		"version": 1,
 		"url": "https://raw.senmanga.com",
 		"contentRating": 1,
-		"languages": ["ja"]
+		"languages": ["ja"],
+		"minAppVersion": "0.7.1"
 	}
 }
