+++ sources/ja.senmanga/src/lib.rs
@@ -12,6 +12,8 @@ use aidoku::{
 };
 
 mod models;
+#[cfg(test)]
+mod test;
 
 use models::*;
 
@@ -49,7 +51,7 @@ impl Source for SenManga {
 		let mut params = QueryParameters::new();
 		params.push("page", Some(&page.to_string()));
 
-		if let Some(query) = query.as_deref().filter(|query| !query.is_empty()) {
+		if let Some(query) = query.as_deref() {
 			params.push("query", Some(query));
 		}
 
@@ -143,7 +145,12 @@ impl Source for SenManga {
 				Some("Manhwa") | Some("Manhua") => Viewer::Webtoon,
 				_ => Viewer::RightToLeft,
 			};
-			manga.status = parse_status(status.as_deref());
+			// This endpoint answers with a null status, while the directory
+			// entry the manga came from carries one, so the existing value is
+			// kept rather than overwritten with Unknown.
+			if let Some(status) = status {
+				manga.status = parse_status(Some(&status));
+			}
 			manga.url = Some(format!("{BASE_URL}/manga/{}/", manga.key));
 			manga.title = title;
 			manga.cover = cover;
@@ -268,153 +275,3 @@ fn content_rating(tags: &[String]) -> ContentRating {
 }
 
 register_source!(SenManga, DeepLinkHandler);
-
-#[cfg(test)]
-mod test {
-	use super::*;
-	use aidoku::alloc::vec;
-	use aidoku_test::aidoku_test;
-
-	/// A long-running series, so its chapter list stays large enough to assert on.
-	const TEST_MANGA_KEY: &str = "one-piece";
-
-	#[aidoku_test]
-	fn browse_test() {
-		let source = SenManga;
-		let result = source
-			.get_search_manga_list(None, 1, vec![])
-			.expect("get_search_manga_list failed");
-
-		assert!(result.entries.len() >= 10);
-		// The directory spans hundreds of pages, so the first one is never last.
-		assert!(result.has_next_page);
-
-		for manga in result.entries {
-			assert!(!manga.key.is_empty());
-			assert!(!manga.title.is_empty());
-			assert!(
-				manga
-					.cover
-					.as_deref()
-					.is_some_and(|cover| cover.starts_with("https://"))
-			);
-		}
-	}
-
-	#[aidoku_test]
-	fn search_test() {
-		let source = SenManga;
-
-		// The api matches against alternative titles too, so a japanese query
-		// has to reach the romanized entry.
-		let result = source
-			.get_search_manga_list(Some("ワンピース".into()), 1, vec![])
-			.expect("get_search_manga_list failed");
-
-		assert!(!result.entries.is_empty());
-		assert!(
-			result
-				.entries
-				.iter()
-				.any(|manga| manga.key == TEST_MANGA_KEY)
-		);
-	}
-
-	#[aidoku_test]
-	fn manga_details_test() {
-		let source = SenManga;
-		let manga = source
-			.get_manga_update(
-				Manga {
-					key: TEST_MANGA_KEY.into(),
-					..Default::default()
-				},
-				true,
-				true,
-			)
-			.expect("get_manga_update failed");
-
-		assert!(!manga.title.is_empty());
-		assert!(manga.cover.is_some());
-		assert!(manga.description.is_some());
-		assert!(manga.tags.as_ref().is_some_and(|tags| !tags.is_empty()));
-		assert_eq!(manga.viewer, Viewer::RightToLeft);
-
-		let chapters = manga.chapters.expect("no chapters");
-		assert!(chapters.len() >= 100);
-
-		for chapter in chapters {
-			assert!(!chapter.key.is_empty());
-			assert!(chapter.chapter_number.is_some());
-			// Setting a language would hide every chapter behind the app's
-			// language filter, since this source is japanese-only.
-			assert!(chapter.language.is_none());
-		}
-	}
-
-	#[aidoku_test]
-	fn page_list_test() {
-		let source = SenManga;
-		let manga = source
-			.get_manga_update(
-				Manga {
-					key: TEST_MANGA_KEY.into(),
-					..Default::default()
-				},
-				false,
-				true,
-			)
-			.expect("get_manga_update failed");
-		let chapter = manga
-			.chapters
-			.as_ref()
-			.and_then(|chapters| chapters.first())
-			.expect("no chapters")
-			.clone();
-
-		let pages = source
-			.get_page_list(manga, chapter)
-			.expect("get_page_list failed");
-
-		assert!(!pages.is_empty());
-		for page in pages {
-			match page.content {
-				PageContent::Url(url, _) => assert!(url.starts_with("https://")),
-				_ => panic!("expected a url page"),
-			}
-		}
-	}
-
-	#[aidoku_test]
-	fn deep_link_test() {
-		let source = SenManga;
-
-		let result = source
-			.handle_deep_link(format!("{BASE_URL}/manga/{TEST_MANGA_KEY}/"))
-			.expect("handle_deep_link failed");
-		assert_eq!(
-			result,
-			Some(DeepLinkResult::Manga {
-				key: TEST_MANGA_KEY.into()
-			})
-		);
-
-		let result = source
-			.handle_deep_link(format!(
-				"{BASE_URL}/manga/{TEST_MANGA_KEY}/chapter-8.338323/"
-			))
-			.expect("handle_deep_link failed");
-		assert_eq!(
-			result,
-			Some(DeepLinkResult::Chapter {
-				manga_key: TEST_MANGA_KEY.into(),
-				key: "8.338323".into()
-			})
-		);
-
-		let result = source
-			.handle_deep_link(format!("{BASE_URL}/directory"))
-			.expect("handle_deep_link failed");
-		assert_eq!(result, None);
-	}
-}
